+++ src/ai/heuristic/Heuristic.java
@@ -3,5 +3,5 @@ package ai.heuristic;
 import ai.state.State;
 
 public interface Heuristic {
-    int evaluate(State state);
+    double evaluate(State state);
 }
